refactor(Modal&Navbar): create subdirectories for modal and navbar related files
