+++ root/frontend/react-src/src/common/modal/modalActions.js
@@ -1,4 +1,4 @@
-import { SHOW_MODAL, HIDE_MODAL } from '../constants/actionTypes'
+import { SHOW_MODAL, HIDE_MODAL } from '../../constants/actionTypes'
 
 // modalType is a string
 export function showModal(modalType, modalProps = {}) {
