+++ root/frontend/react-src/src/common/navbar/Navbar.css
